feature: review 유저 객체 매핑

+++ src/main/java/com/reallyeasy/cineView/domain/review/service/ReviewService.java
@@ -4,22 +4,35 @@ import com.reallyeasy.cineView.domain.review.dto.request.ReviewRequest;
 import com.reallyeasy.cineView.domain.review.dto.response.ReviewResponse;
 import com.reallyeasy.cineView.domain.review.entity.Review;
 import com.reallyeasy.cineView.domain.review.repository.ReviewRepository;
+import com.reallyeasy.cineView.domain.user.entity.User;
+import com.reallyeasy.cineView.domain.user.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@Transactional
 @AllArgsConstructor
 public class ReviewService {
     private final ReviewRepository reviewRepository;
+    private final UserRepository userRepository;
 
     public ReviewResponse createReview(ReviewRequest request, Long userId, Long movieId) {
-        Review review = reviewRepository.save(new Review(request, userId, movieId));
+        User user = userRepository.findById(userId).orElseThrow();
+        Review review = reviewRepository.save(new Review(request, user, movieId));
         return new ReviewResponse(review);
     }
 
     public ReviewResponse updateReview(ReviewRequest request, Long userId, Long movieId, Long reviewId) {
         Review review = reviewRepository.findById(reviewId).orElseThrow();
-        // todo movie, user
+        User user = userRepository.findById(userId).orElseThrow();
+        if (!review.getUser().equals(user))
+            throw new IllegalArgumentException("해당 리뷰를 수정할 권한이 없습니다.");
+
+        // todo: 영화 ID 일치 여부 확인
+        if (!review.getMovieId().equals(movieId))
+            throw new IllegalArgumentException("리뷰와 영화 정보가 일치하지 않습니다.");
+
         review.updateContent(request.getContent());
         review.updateRating(request.getRating());
 
@@ -28,7 +41,14 @@ public class ReviewService {
 
     public void deleteReview(Long userId, Long movieId, Long reviewId) {
         Review review = reviewRepository.findById(reviewId).orElseThrow();
-        // todo movie, user
+        User user = userRepository.findById(userId).orElseThrow();
+        if (!review.getUser().equals(user))
+            throw new IllegalArgumentException("해당 리뷰를 수정할 권한이 없습니다.");
+
+        // todo: 영화 ID 일치 여부 확인
+        if (!review.getMovieId().equals(movieId))
+            throw new IllegalArgumentException("리뷰와 영화 정보가 일치하지 않습니다.");
+
         reviewRepository.delete(review);
     }
 }
